Started created the main hero section. Made a content.json for better data management. Added more TODO's so I don't forget
