+++ src/assets/data/content.json
@@ -1,0 +1,23 @@
+{
+  "header": {
+    "menuItems": [
+      { "title": "home", "url": "#" },
+      { "title": "shop", "url": "#" },
+      { "title": "about", "url": "#" },
+      { "title": "contact", "url": "#" }
+    ]
+  },
+  "main": {
+    "hero": {
+      "title": "Discover innovative ways to decorate",
+      "description": "We provide unmatched quality, comfort, and style for property owners across the country. Our experts combine form and function in bringing your vision to life. Create a room in your own style with our collection and make your property a reflection of you and what you love.",
+      "buttonText": "shop now",
+      "image": "src/assets/images/image-about-dark.jpg"
+    },
+    "features": {
+      "title": "about our furniture",
+      "description": "Our multifunctional collection blends design and function to suit your individual taste. Make each room unique, or pick a cohesive theme that best express your interests and what inspires you. Find the furniture pieces you need, from traditional to contemporary styles or anything in between. Product specialists are available to help you create your dream space.",
+      "image": "src/assets/images/image-about-light.jpg"
+    }
+  }
+}
