+++ src/components/DesktopMenu.vue
@@ -1,10 +1,10 @@
 <template>
   <ul class="gap-8 text-xl text-white">
-    <li v-for="item in menuItems" :key="item">
+    <li v-for="item in menuItems" :key="item.title">
       <a
-        href="#"
+        :href="item.url"
         class="relative after:absolute after:-bottom-1 after:left-0 after:h-0.5 after:w-0 after:bg-white after:transition-all after:duration-200 hover:after:w-full"
-        >{{ item }}</a
+        >{{ item.title }}</a
       >
     </li>
   </ul>
@@ -12,6 +12,6 @@
 
 <script setup lang="ts">
 defineProps<{
-  menuItems: string[];
+  menuItems: { title: string; url: string }[];
 }>();
 </script>
